chore: add comments and organize code
